refactor: extract file change detection to shared utility
Move duplicate detectChangedFiles logic to src/core/utils/file-change-detection.ts
- Consolidates ~150 lines of duplicate code from parse-typescript-project.tool.ts and incremental-parse.handler.ts
- Adds configurable logWarnings option for different use cases
- Maintains security checks for symlink traversal
